+++ tla-io/src/main/scala/at/forsyte/apalache/io/quint/QuintIR.scala
@@ -24,8 +24,12 @@ private[quint] object QuintDeserializer extends upickle.AttributeTagged {
 
 import QuintDeserializer.{macroRW, ReadWriter => RW}
 
+// Internal error
 private[quint] class QuintIRParseError(errMsg: String)
     extends Exception("Input was not a valid representation of the QuintIR: " + errMsg)
+
+// User facing error
+class QuintUnsupportedError(errMsg: String) extends Exception("Unsupported quint input: " + errMsg)
 
 /** The JSON output produced by quint parse */
 private[quint] case class QuintOutput(
@@ -116,10 +120,17 @@ private[quint] object QuintEx {
     implicit val rw: RW[QuintApp] = macroRW
   }
 
+  case class QuintLambdaParameter(
+      id: Int,
+      name: String)
+  object QuintLambdaParameter {
+    implicit val rw: RW[QuintLambdaParameter] = macroRW
+  }
+
   @key("lambda") case class QuintLambda(
       id: Int,
       /** Identifiers for the formal parameters */
-      params: Seq[String],
+      params: Seq[QuintLambdaParameter],
       /** The qualifier for the defined operator */
       // TODO should this eventually be a sumtype?
       qualifier: String,
@@ -336,11 +347,23 @@ private[quint] object QuintType {
   @key("tup") case class QuintTupleT(fields: Row) extends QuintType
   object QuintTupleT {
     implicit val rw: RW[QuintTupleT] = macroRW
+
+    // Helper for manually constructing tuple types
+    def ofTypes(types: QuintType*): QuintTupleT = {
+      val fields = types.zipWithIndex.map { case (t, i) => RecordField(i.toString, t) }
+      QuintTupleT(Row.Cell(fields, Row.Nil()))
+    }
   }
 
   @key("rec") case class QuintRecordT(fields: Row) extends QuintType
   object QuintRecordT {
     implicit val rw: RW[QuintRecordT] = macroRW
+
+    // Helper for manually constructing record type
+    def ofFieldTypes(fieldTypes: (String, QuintType)*): QuintRecordT = {
+      val fields = fieldTypes.map { case (f, t) => RecordField(f, t) }
+      QuintRecordT(Row.Cell(fields, Row.Nil()))
+    }
   }
 
   case class UnionRecord(tagValue: String, fields: Row)